requirements.txt
fastapi==0.110.0
slack-sdk==3.27.1
openai==1.64.0
pydantic==2.10.6
python-multipart==0.0.6
requests==2.31.0
uvicorn==0.29.0
matplotlib
pandas
pillow